+++ common/src/main/java/common/peer/JoinPeer.java
@@ -6,15 +6,15 @@ import se.sics.kompics.Event;
 
 public class JoinPeer extends Event {
 
-	private final BigInteger peerId;
+	private final Long peerId;
 
 //-------------------------------------------------------------------
-	public JoinPeer(BigInteger peerId) {
+	public JoinPeer(Long peerId) {
 		this.peerId = peerId;
 	}
 
 //-------------------------------------------------------------------
-	public BigInteger getPeerId() {
+	public Long getPeerId() {
 		return this.peerId;
 	}
 }
