+++ common/src/main/java/common/simulation/AddIndexEntry.java
@@ -6,13 +6,13 @@ import se.sics.kompics.Event;
 
 public final class AddIndexEntry extends Event {
 
-    private final BigInteger id;
+    private final Long id;
 
-    public AddIndexEntry(BigInteger id) {
+    public AddIndexEntry(Long id) {
         this.id = id;
     }
 
-    public BigInteger getId() {
+    public Long getId() {
         return id;
     }
