+++ common/src/main/java/common/simulation/scenarios/Operations.java
@@ -11,18 +11,18 @@ import se.sics.kompics.p2p.experiment.dsl.adaptor.Operation1;
 @SuppressWarnings("serial")
 public class Operations {
 
-  	public static Operation1<AddIndexEntry, BigInteger> addIndexEntry() {
-		return new Operation1<AddIndexEntry, BigInteger>() {
+  	public static Operation1<AddIndexEntry, Long> addIndexEntry() {
+		return new Operation1<AddIndexEntry, Long>() {
                         @Override
-			public AddIndexEntry generate(BigInteger id) {
+			public AddIndexEntry generate(Long id) {
 				return new AddIndexEntry(id);
 			}
 		};
 	}
 //-------------------------------------------------------------------
-	public static Operation1<PeerJoin, BigInteger> peerJoin(final int num) {
-		return new Operation1<PeerJoin, BigInteger>() {
-			public PeerJoin generate(BigInteger id) {
+	public static Operation1<PeerJoin, Long> peerJoin(final int num) {
+		return new Operation1<PeerJoin, Long>() {
+			public PeerJoin generate(Long id) {
 				return new PeerJoin(id, num);
 			}
 		};
@@ -31,7 +31,7 @@ public class Operations {
 //-------------------------------------------------------------------
 	public static Operation1<PeerFail, Long> peerFail = new Operation1<PeerFail, Long>() {
 		public PeerFail generate(Long id) {
-			return new PeerFail(BigInteger.valueOf(id));
+			return new PeerFail(id);
 		}
 	};
